@@ -2,7 +2,7 @@
 name = "wheat-from-chaff-backend"
 version = "0.1.0"
 description = "FastAPI backend for the candidate filtering system."
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 dependencies = [
     "fastapi",
     "uvicorn[standard]",
@@ -15,6 +15,8 @@ dependencies = [
     "azure-identity",
     "pdf2image",
     "Pillow",
+    "loguru",
+    "vllm>=0.10.0",
 ]
 
 [project.scripts]
